Refactoring IntCodeProgram to support long data values

+++ Common/Streams/IInputStream.cs
@@ -2,6 +2,6 @@ namespace Common.Streams
 {
     public interface IInputStream
     {
-        int Read();
+        long Read();
     }
 }
