+++ Common/Streams/InputStream.cs
@@ -2,15 +2,15 @@ namespace Common.Streams
 {
     public class InputStream : IInputStream
     {
-        private readonly List<int> input;
+        private readonly List<long> input;
         private int pos = 0;
         
-        public InputStream(IEnumerable<int> input)
+        public InputStream(IEnumerable<long> input)
         {
             this.input = input.ToList();
         }
 
-        public int Read()
+        public long Read()
         {
             return input[pos++];
         }
